+++ packages/providers/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "providers/5.0.0-beta.139";
+export declare const version = "providers/5.0.0-beta.140";