+++ packages/providers/fallback-provider.js
@@ -138,7 +138,9 @@ var FallbackProvider = /** @class */ (function (_super) {
             case "getStorageAt":
                 return provider.getStorageAt(params.address, params.position, params.blockTag || "latest");
             case "sendTransaction":
-                return provider.sendTransaction(params.signedTransaction);
+                return provider.sendTransaction(params.signedTransaction).then(function (result) {
+                    return result.hash;
+                });
             case "getBlock":
                 return provider[(params.includeTransactions ? "getBlockWithTransactions" : "getBlock")](params.blockTag || params.blockHash);
             case "call":
@@ -253,10 +255,44 @@ var FallbackProvider = /** @class */ (function (_super) {
                     }
                     // Out of options; give up
                     if (runners.length === 0 && inflightWeight === 0) {
+                        // @TODO: this might need some more thinking... Maybe only if half
+                        // of the results contain non-error?
+                        if (method === "getGasPrice") {
+                            var values_1 = [];
+                            Object.keys(results).forEach(function (key) {
+                                results[key].forEach(function (result) {
+                                    if (!result.result) {
+                                        return;
+                                    }
+                                    values_1.push(result.result);
+                                });
+                            });
+                            values_1.sort(function (a, b) {
+                                if (a.lt(b)) {
+                                    return -1;
+                                }
+                                if (a.gt(b)) {
+                                    return 1;
+                                }
+                                return 0;
+                            });
+                            var index = parseInt(String(values_1.length / 2));
+                            if (values_1.length % 2) {
+                                resolve(values_1[index]);
+                                return;
+                            }
+                            resolve(values_1[index - 1].add(values_1[index]).div(2));
+                            return;
+                        }
                         if (firstError === null) {
                             firstError = logger.makeError("failed to meet quorum", logger_1.Logger.errors.SERVER_ERROR, {
                                 results: Object.keys(results).map(function (u) {
-                                    return { result: u, weight: results[u].reduce(function (accum, r) { return (accum + r.weight); }, 0) };
+                                    return {
+                                        method: method,
+                                        params: params,
+                                        result: u,
+                                        weight: results[u].reduce(function (accum, r) { return (accum + r.weight); }, 0)
+                                    };
                                 })
                             });
                         }